Create .gitignore
Okay, so I created a new directory in my local Repository dir named "ignore" (and two subdirs called "evaluation" and "letters". It contains all the extracted letters and the evaluations from last time and the paths in the code use the folders, so in order to use the code I wrote as it is, you have to create them local as well. I put the "ignore" directory on the gitignore list so we don't have to worry about uploading too much stuff^^